Moved common rune fields into bestiary.Rune, changed bestiary.Rune to abstract model
