+++ herders/forms.py
@@ -8,6 +8,7 @@ from dal import autocomplete
 from django import forms
 from django.contrib.auth.forms import AuthenticationForm, PasswordChangeForm, PasswordResetForm, SetPasswordForm
 from django.contrib.auth.models import User
+from django.contrib.postgres.forms import SplitArrayField
 from django.core.validators import RegexValidator
 from django.db.models.fields import BLANK_CHOICE_DASH
 from django.forms import ModelForm
@@ -898,18 +899,18 @@ class AddRuneInstanceForm(ModelForm):
         self.fields['main_stat_value'].label = False
         self.fields['innate_stat'].label = False
         self.fields['innate_stat_value'].label = False
-        self.fields['substat_1'].label = False
-        self.fields['substat_1_value'].label = False
-        self.fields['substat_1_craft'].label = False
-        self.fields['substat_2'].label = False
-        self.fields['substat_2_value'].label = False
-        self.fields['substat_2_craft'].label = False
-        self.fields['substat_3'].label = False
-        self.fields['substat_3_value'].label = False
-        self.fields['substat_3_craft'].label = False
-        self.fields['substat_4'].label = False
-        self.fields['substat_4_value'].label = False
-        self.fields['substat_4_craft'].label = False
+        # self.fields['substat_1'].label = False
+        # self.fields['substat_1_value'].label = False
+        # self.fields['substat_1_craft'].label = False
+        # self.fields['substat_2'].label = False
+        # self.fields['substat_2_value'].label = False
+        # self.fields['substat_2_craft'].label = False
+        # self.fields['substat_3'].label = False
+        # self.fields['substat_3_value'].label = False
+        # self.fields['substat_3_craft'].label = False
+        # self.fields['substat_4'].label = False
+        # self.fields['substat_4_value'].label = False
+        # self.fields['substat_4_craft'].label = False
         self.fields['assigned_to'].label = False
         self.fields['notes'].label = False
 
@@ -945,32 +946,33 @@ class AddRuneInstanceForm(ModelForm):
                     ),
                     Div(
                         HTML('<label class="col-md-2 control-label">Substat 1</label>'),
-                        Field('substat_1', wrapper_class='col-md-4 inline-horizontal'),
-                        Field('substat_1_value', wrapper_class='col-md-3 inline-horizontal', placeholder='Value'),
-                        Field('substat_1_craft', wrapper_class='col-md-3 inline-horizontal'),
+                        Field('substats', wrapper_class='col-md-4 inline-horizontal'),
+                        Field('substat_values', wrapper_class='col-md-3 inline-horizontal', placeholder='Value'),
+                        Field('substat_crafts', wrapper_class='col-md-3 inline-horizontal'),
+                        Field('substat_craft_values', wrapper_class='col-md-3 inline-horizontal'),
                         css_class='form-group form-group-condensed',
                     ),
-                    Div(
-                        HTML('<label class="col-md-2 control-label">Substat 2</label>'),
-                        Field('substat_2', wrapper_class='col-md-4 inline-horizontal'),
-                        Field('substat_2_value', wrapper_class='col-md-3 inline-horizontal', placeholder='Value'),
-                        Field('substat_2_craft', wrapper_class='col-md-3 inline-horizontal'),
-                        css_class='form-group form-group-condensed',
-                    ),
-                    Div(
-                        HTML('<label class="col-md-2 control-label">Substat 3</label>'),
-                        Field('substat_3', wrapper_class='col-md-4 inline-horizontal'),
-                        Field('substat_3_value', wrapper_class='col-md-3 inline-horizontal', placeholder='Value'),
-                        Field('substat_3_craft', wrapper_class='col-md-3 inline-horizontal'),
-                        css_class='form-group form-group-condensed',
-                    ),
-                    Div(
-                        HTML('<label class="col-md-2 control-label">Substat 4</label>'),
-                        Field('substat_4', wrapper_class='col-md-4 inline-horizontal'),
-                        Field('substat_4_value', wrapper_class='col-md-3 inline-horizontal', placeholder='Value'),
-                        Field('substat_4_craft', wrapper_class='col-md-3 inline-horizontal'),
-                        css_class='form-group form-group-condensed',
-                    ),
+                    # Div(
+                    #     HTML('<label class="col-md-2 control-label">Substat 2</label>'),
+                    #     Field('substat_2', wrapper_class='col-md-4 inline-horizontal'),
+                    #     Field('substat_2_value', wrapper_class='col-md-3 inline-horizontal', placeholder='Value'),
+                    #     Field('substat_2_craft', wrapper_class='col-md-3 inline-horizontal'),
+                    #     css_class='form-group form-group-condensed',
+                    # ),
+                    # Div(
+                    #     HTML('<label class="col-md-2 control-label">Substat 3</label>'),
+                    #     Field('substat_3', wrapper_class='col-md-4 inline-horizontal'),
+                    #     Field('substat_3_value', wrapper_class='col-md-3 inline-horizontal', placeholder='Value'),
+                    #     Field('substat_3_craft', wrapper_class='col-md-3 inline-horizontal'),
+                    #     css_class='form-group form-group-condensed',
+                    # ),
+                    # Div(
+                    #     HTML('<label class="col-md-2 control-label">Substat 4</label>'),
+                    #     Field('substat_4', wrapper_class='col-md-4 inline-horizontal'),
+                    #     Field('substat_4_value', wrapper_class='col-md-3 inline-horizontal', placeholder='Value'),
+                    #     Field('substat_4_craft', wrapper_class='col-md-3 inline-horizontal'),
+                    #     css_class='form-group form-group-condensed',
+                    # ),
                     Div(
                         HTML('<label class="col-md-2 control-label">Assign To</label>'),
                         Div(
@@ -998,14 +1000,23 @@ class AddRuneInstanceForm(ModelForm):
 
     class Meta:
         model = RuneInstance
+        # field_classes = {
+        #     'substats': SplitArrayField(
+        #         base_field=forms.ChoiceField(required=False, choices=RuneInstance.STAT_CHOICES),
+        #         size=4,
+        #         remove_trailing_nulls=True,
+        #     ),
+        #     'substat_values': SplitArrayField(
+        #         base_field=forms.IntegerField,
+        #         size=4,
+        #         remove_trailing_nulls=True,
+        #     )
+        # }
         fields = (
             'type', 'stars', 'level', 'slot',
             'main_stat', 'main_stat_value',
             'innate_stat', 'innate_stat_value',
-            'substat_1', 'substat_1_value', 'substat_1_craft',
-            'substat_2', 'substat_2_value', 'substat_2_craft',
-            'substat_3', 'substat_3_value', 'substat_3_craft',
-            'substat_4', 'substat_4_value', 'substat_4_craft',
+            'substats', 'substat_values', 'substat_crafts', 'substat_craft_values',
             'assigned_to', 'notes', 'marked_for_sale',
         )
         widgets = {
